Split ColorizerPalette into engine and interface
